refactor: move counter to its own node
