+++ CMakeLists.txt
@@ -160,7 +160,7 @@ include_directories(
 ## Mark executable scripts (Python etc.) for installation
 ## in contrast to setup.py, you can choose the destination
 catkin_install_python(
-  PROGRAMS scripts/heartbeat.py scripts/followers.py
+  PROGRAMS scripts/heartbeat.py scripts/followers.py scripts/collisions_counter.py
   DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
