+++ launch/followers.launch
@@ -2,4 +2,5 @@
     <node pkg="turtlesim" type="turtlesim_node" name="sim"/>
     <node pkg="turtlesim" type="turtle_teleop_key" name="teleop" output="screen"/>
     <node pkg="follower3" type="followers.py" name="followers"/>
+    <node pkg="follower3" type="collisions_counter.py" name="collisions_counter"/>
 </launch>
